outline_builder/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
cycler==0.12.1
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imutils==0.5.4
isodate==0.7.2
kiwisolver==1.4.8
numpy==2.2.1
ollama==0.4.7
opencv-python==4.10.0.84
packaging==24.2
pillow==11.1.0
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
